Move positions table to own component, stub out Overview stats
