+++ apps/fixed-borrow/src/ui/portfolio/PortfolioPage/PortfolioPage.tsx
@@ -1,138 +1,27 @@
 import { ReactElement } from "react";
-import { Address, useAccount, useToken } from "wagmi";
-import { SortableGridTable } from "src/ui/base/tables/SortableGridTable";
-import { useUserLoans } from "src/ui/loans/hooks/useUserLoans";
-import { Hash } from "@wagmi/core";
-import { useTransactionTimestamp } from "src/ui/base/transactions/useTransactionTimestamp/useTransactionTimestamp";
-import { formatBigInt } from "src/base/bigint/formatBigInt";
-import { formatBalance } from "src/ui/base/formatting/formatBalance";
-import { AssetIcon } from "src/ui/token/AssetIcon";
+import { StatWell } from "src/ui/base/StatWell/StatWell";
+import { PositionsTable } from "src/ui/portfolio/PositionsTable/PositionsTable";
 
 export function PortfolioPage(): ReactElement {
-  const { address: account } = useAccount();
-
-  const { userLoans = [] } = useUserLoans(account);
   return (
-    <div className="flex flex-col gap-3">
-      <h4 className="h4 font-bold text-white">Positions</h4>
-      <div className="max-w-4xl">
-        <SortableGridTable
-          headingRowClassName="grid-cols-[1fr_1fr_1fr_1fr_64px] gap-10"
-          bodyRowClassName="group grid-cols-[1fr_1fr_1fr_1fr_64px] gap-10"
-          emptyTableElement={<span className="text-white">No loans found</span>}
-          cols={[
-            {
-              cell: <span className="text-secondaryText">Date opened</span>,
-              sortKey: "openDate",
-            },
-            {
-              cell: <span className="text-secondaryText">Rate</span>,
-              sortKey: "votingPower",
-            },
-            {
-              cell: <span className="text-secondaryText">Collateral</span>,
-              sortKey: "collateralAmount",
-            },
-            {
-              cell: <span className="text-secondaryText">Debt</span>,
-              sortKey: "debtAmount",
-            },
-          ]}
-          rows={[...userLoans, ...userLoans, ...userLoans].map(
-            (
-              {
-                txHash,
-                shortId,
-                collateralDeposited,
-                collateralTokenAddress,
-                borrowedAmount,
-                borrowTokenAddress,
-              },
-              i,
-            ) => ({
-              cells: [
-                <DateOpenedCell
-                  key={`dateopened-${txHash}-${i}`}
-                  txHash={txHash}
-                />,
-                // TODO: Implement this
-                <RateCell
-                  key={`rate-${txHash}-${i}`}
-                  shortId={shortId}
-                  txHash={txHash}
-                />,
-                <AmountCell
-                  key={`collateral-${txHash}-${i}`}
-                  amount={collateralDeposited}
-                  tokenAddress={collateralTokenAddress}
-                />,
-                <AmountCell
-                  key={`debt-${txHash}-${i}`}
-                  amount={borrowedAmount}
-                  tokenAddress={borrowTokenAddress}
-                />,
-                <div
-                  key={`expand-${txHash}-${i}`}
-                  className="flex shrink-0 items-center"
-                >
-                  <img src="/caret-down.svg" />
-                </div>,
-              ],
-            }),
-          )}
-        />
+    <div className="flex flex-col gap-20">
+      <div className="space-y-3">
+        <h4 className="font-bold text-white">Overview</h4>
+        <div className="max-w-4xl">
+          <div className="flex gap-8 ">
+            <StatWell label="Collateral" stat="$130,491.20" />
+            <StatWell label="Total debt" stat="$9,824.62" />
+            <StatWell label="Borrow limit" stat="$83,218.81" />
+            <StatWell label="Loan to value" stat="7.5%" />
+          </div>
+        </div>
       </div>
-    </div>
-  );
-}
-
-function DateOpenedCell({ txHash }: { txHash: Hash }): ReactElement {
-  const { timestamp } = useTransactionTimestamp(txHash);
-  return (
-    <span className="text-lg text-white">
-      {timestamp ? new Date(timestamp).toDateString() : ""}
-    </span>
-  );
-}
-
-const SIX_MONTHS_IN_MILLISECONDS = 15778800000;
-function RateCell({ txHash }: { shortId: bigint; txHash: Hash }): ReactElement {
-  const { timestamp } = useTransactionTimestamp(txHash);
-  return (
-    <div className="flex flex-col ">
-      <span className="inline-flex items-center gap-2 text-h6 text-white">
-        <span>1.50%</span>
-        <span className="text-success">Fixed</span>
-      </span>
-      <span className="leading-sm text-secondaryText">
-        {timestamp
-          ? `until ${new Date(
-              timestamp + SIX_MONTHS_IN_MILLISECONDS,
-            ).toDateString()}`
-          : ""}
-      </span>
-    </div>
-  );
-}
-
-function AmountCell({
-  amount,
-  tokenAddress,
-}: {
-  amount: bigint;
-  tokenAddress: Address;
-}): ReactElement {
-  const { data: tokenMetadata } = useToken({ address: tokenAddress });
-  const amountLabel = formatBalance(
-    formatBigInt(amount, tokenMetadata?.decimals),
-  );
-  return (
-    <div className="flex flex-col ">
-      <span className="inline-flex items-center gap-1 text-h6 text-white">
-        <AssetIcon size="sm" address={tokenAddress} /> {amountLabel}{" "}
-        {tokenMetadata?.symbol}
-      </span>
-      <span className="leading-sm text-secondaryText">$2,302.13</span>
+      <div className="space-y-2">
+        <h4 className="font-bold text-white">Positions</h4>
+        <div className="max-w-4xl">
+          <PositionsTable />
+        </div>
+      </div>
     </div>
   );
 }
